+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Head
 
+- [Breaking] Use ES2015 module syntax in default template.
 - [Breaking] Rename `modules` option to `prependJs`.
 
 ## 0.6.0